+++ CopilotKit/packages/runtime/src/lib/runtime/remote-lg-action.ts
@@ -161,7 +161,13 @@ async function streamEvents(controller: ReadableStreamDefaultController, args: E
     payload.command = { resume: formattedMessages[formattedMessages.length - 1] };
   }
   if (lgInterruptMetaEvent?.response) {
-    payload.command = { resume: lgInterruptMetaEvent.response };
+    let response = lgInterruptMetaEvent.response;
+    try {
+      payload.command = { resume: JSON.parse(response) };
+      // In case of unparsable string, we keep the event as is
+    } catch (e) {
+      payload.command = { resume: response };
+    }
   }
 
   if (mode === "continue" && !activeInterruptEvent) {
@@ -234,11 +240,12 @@ async function streamEvents(controller: ReadableStreamDefaultController, args: E
           typeof interruptValue != "string" &&
           "__copilotkit_interrupt_value__" in interruptValue
         ) {
+          const evValue = interruptValue.__copilotkit_interrupt_value__;
           emit(
             JSON.stringify({
               event: LangGraphEventTypes.OnCopilotKitInterrupt,
               data: {
-                value: interruptValue.__copilotkit_interrupt_value__,
+                value: typeof evValue === "string" ? evValue : JSON.stringify(evValue),
                 messages: langchainMessagesToCopilotKit(interruptValue.__copilotkit_messages__),
               },
             }) + "\n",
@@ -247,7 +254,10 @@ async function streamEvents(controller: ReadableStreamDefaultController, args: E
           emit(
             JSON.stringify({
               event: LangGraphEventTypes.OnInterrupt,
-              value: interruptValue,
+              value:
+                typeof interruptValue === "string"
+                  ? interruptValue
+                  : JSON.stringify(interruptValue),
             }) + "\n",
           );
         }
@@ -261,7 +271,7 @@ async function streamEvents(controller: ReadableStreamDefaultController, args: E
       }
 
       const event = chunk.data;
-      const currentNodeName = event.name;
+      const currentNodeName = event.metadata.langgraph_node;
       const eventType = event.event;
       const runId = event.metadata.run_id;
       externalRunId = runId;