+++ examples/coagents-starter/ui/app/api/copilotkit/route.ts
@@ -1,14 +1,26 @@
 import { NextRequest } from "next/server";
 import {
   CopilotRuntime,
-  OpenAIAdapter,
   copilotRuntimeNextJSAppRouterEndpoint,
+  ExperimentalEmptyAdapter,
+  // langGraphPlatformEndpoint
 } from "@copilotkit/runtime";
 
-const serviceAdapter = new OpenAIAdapter({});
+const serviceAdapter = new ExperimentalEmptyAdapter();
 
 const runtime = new CopilotRuntime({
   remoteEndpoints: [
+    // Uncomment this if you want to use LangGraph JS, make sure to
+    // remove the remote action url below too.
+    //
+    // langGraphPlatformEndpoint({
+    //   deploymentUrl: "http://localhost:8123",
+    //   langsmithApiKey: process.env.LANGSMITH_API_KEY || "", // only used in LangGraph Platform deployments
+    //   agents: [{
+    //       name: 'sample_agent',
+    //       description: 'A helpful LLM agent.'
+    //   }]
+    // }),
     {
       url: process.env.REMOTE_ACTION_URL || "http://localhost:8000/copilotkit",
     },