+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,87 @@
 # @copilotkit/shared
 
+## 1.5.19
+
+### Patch Changes
+
+- 0dd1ab9: - fix(errors): allow non copilotkit errors to pass to consumer app error boundary
+
+## 1.5.19-next.1
+
+### Patch Changes
+
+- 0dd1ab9: - fix(errors): allow non copilotkit errors to pass to consumer app error boundary
+
+## 1.5.19-next.0
+
+## 1.5.18
+
+### Patch Changes
+
+- d47cd26: - fix: detect and alert on version mismatch
+- f77a7b9: - fix: use warning when version mismatch is not expected to error out
+- 38d3ac2: - fix: add additional info the our error messages
+
+## 1.5.18-next.3
+
+### Patch Changes
+
+- f77a7b9: - fix: use warning when version mismatch is not expected to error out
+
+## 1.5.18-next.2
+
+### Patch Changes
+
+- 38d3ac2: - fix: add additional info the our error messages
+
+## 1.5.18-next.1
+
+## 1.5.18-next.0
+
+### Patch Changes
+
+- d47cd26: - fix: detect and alert on version mismatch
+
+## 1.5.17
+
+### Patch Changes
+
+- 1fc3902: - Revert "fix: detect and alert on version mismatch (#1333)"
+
+  This reverts commit 48b7c7b1bd48ced82ffb9a00d6eddc1f7581e0c1.
+
+## 1.5.17-next.0
+
+### Patch Changes
+
+- 1fc3902: - Revert "fix: detect and alert on version mismatch (#1333)"
+
+  This reverts commit 48b7c7b1bd48ced82ffb9a00d6eddc1f7581e0c1.
+
+## 1.5.16
+
+### Patch Changes
+
+- 48b7c7b: - fix: detect and alert on version mismatch
+
+## 1.5.16-next.2
+
+## 1.5.16-next.1
+
+### Patch Changes
+
+- 48b7c7b: - fix: detect and alert on version mismatch
+
+## 1.5.16-next.0
+
+## 1.5.15
+
+### Patch Changes
+
+- 7b3141d: - feat(interrupt): support LG interrupt with useLangGraphInterrupt hook
+  - chore(interrupt): add e2e test to interrupt functionality
+  - feat(interrupt): add support for multiple interrupts and conditions
+
 ## 1.5.15-next.8
 
 ## 1.5.15-next.7